+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+**New decoders:**
+
+- `startsWith(prefix)`
+- `endsWith(suffix)`
+
 ## [2.4.2] - 2024-06-30
 
 - Fix a regression in `taggedUnion` (thanks for reporting, @programever)